refactor(worker): extract agent-turn credentials, tools, and agent build
Move git/codemode/run-credential mint and renewal into
prepareRunCredentials, tool policy/Variable Sets/MCP/prepareTools into
prepareTurnToolPolicy and prepareTurnToolRuntime, and agent
construction plus image-reference resolution into buildTurnAgent.
